Refactor authentication middleware; consolidate logic and enhance user redirection for protected routes
