--- middleware.tsx
@@ -1,6 +1,0 @@
-export { auth as middleware } from "@/auth"
-
-export const config = {
-    runtime: "nodejs",
-    matcher: ["/((?!api|_next/static|_next/image|favicon.ico).*)"],
-}
